# src/gen/java/io/swagger/model/EddList.java
/*
 * Expected Delivery Date
 * This is an api for getting expected delivery date for a pregnant women.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: aakash.verma@cerner.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package io.swagger.model;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.Objects;

import com.fasterxml.jackson.annotation.JsonProperty;

import io.swagger.annotations.ApiModelProperty;

/**
 * EddList
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2019-09-08T11:58:17.132Z")
public class EddList {
	@JsonProperty("userId")
	private Long userId = null;

	@JsonProperty("isEnabled")
	private Boolean isEnabled = null;

	@JsonProperty("eddMapList")
	private HashMap<Integer, ArrayList<HashMap<String, Object>>> eddMapList = null;

//	@JsonProperty("IVF")
//	private HashMap<Integer, ArrayList<HashMap<String, Object>>> IVF = null;
//
//	@JsonProperty("LMP")
//	private HashMap<Integer, ArrayList<HashMap<String, Object>>> LMP = null;

	@JsonProperty("latestEDD")
	private Edd latestEDD = null;

	public EddList userId(Long userId) {
		this.userId = userId;
		return this;
	}

	/**
	 * Get userId
	 * 
	 * @return userId
	 **/
	@JsonProperty("userId")
	@ApiModelProperty(value = "")
	public Long getUserId() {
		return userId;
	}

	public void setUserId(Long userId) {
		this.userId = userId;
	}

	
	/**
	 * Get isEnabled
	 * 
	 * @return isEnabled
	 **/
	@JsonProperty("isEnabled")
	@ApiModelProperty(value = "")
	public Boolean getIsEnabled() {
		return isEnabled;
	}

	public void setIsEnabled(Boolean isEnabled) {
		this.isEnabled = isEnabled;
	}
	
	/**
	 * Get userId
	 * 
	 * @return userId
	 **/
	@JsonProperty("latestEDD")
	@ApiModelProperty(value = "")
	public Edd getLatestEDD() {
		return latestEDD;
	}

	public void setLatestEDD(Edd latestEdd) {
		this.latestEDD = latestEdd;
	}

	public EddList eddMapList(HashMap<Integer, ArrayList<HashMap<String, Object>>> eddMapList) {
		this.eddMapList = eddMapList;
		return this;
	}

	public EddList addEddItem(Integer refUserEDDId, HashMap<String, Object> eddMapItem) {
		if (this.eddMapList == null) {
			this.eddMapList = new HashMap<Integer, ArrayList<HashMap<String, Object>>>();
		}
		ArrayList<HashMap<String, Object>> eddList = this.eddMapList.get(refUserEDDId);
		if (eddList == null) {
			eddList = new ArrayList<HashMap<String, Object>>();
		}
		eddList.add(eddMapItem);
		this.eddMapList.put(refUserEDDId, eddList);
		return this;
	}

//	/**
//	 * Get Ultrasound
//	 * 
//	 * @return Ultrasound
//	 **/
//	@JsonProperty("Ultrasound")
//	@ApiModelProperty(value = "")
//	public HashMap<Integer, ArrayList<HashMap<String, Object>>> getUltrasound() {
//		return Ultrasound;
//	}
//
//	public void setUltrasound(HashMap<Integer, ArrayList<HashMap<String, Object>>> Ultrasound) {
//		this.Ultrasound = Ultrasound;
//	}
//
//	public EddList IVF(HashMap<Integer, ArrayList<HashMap<String, Object>>> IVF) {
//		this.IVF = IVF;
//		return this;
//	}
//
//	public EddList addIVFItem(Integer refUserEDDId, HashMap<String, Object> IVFItem) {
//		if (this.IVF == null) {
//			this.IVF = new HashMap<Integer, ArrayList<HashMap<String, Object>>>();
//		}
//		ArrayList<HashMap<String, Object>> IVFList = this.IVF.get(refUserEDDId);
//		if (IVFList == null) {
//			IVFList = new ArrayList<HashMap<String, Object>>();
//		}
//		IVFList.add(IVFItem);
//		this.IVF.put(refUserEDDId, IVFList);
//		return this;
//	}
//
//	/**
//	 * Get IVF
//	 * 
//	 * @return IVF
//	 **/
//	@JsonProperty("IVF")
//	@ApiModelProperty(value = "")
//	public HashMap<Integer, ArrayList<HashMap<String, Object>>> getIVF() {
//		return IVF;
//	}
//
//	public void setIVF(HashMap<Integer, ArrayList<HashMap<String, Object>>> IVF) {
//		this.IVF = IVF;
//	}
//
//	public EddList LMP(HashMap<Integer, ArrayList<HashMap<String, Object>>> LMP) {
//		this.LMP = LMP;
//		return this;
//	}
//
//	public EddList addLMPItem(Integer refUserEDDId, HashMap<String, Object> LMPItem) {
//		if (this.LMP == null) {
//			this.LMP = new HashMap<Integer, ArrayList<HashMap<String, Object>>>();
//		}
//		ArrayList<HashMap<String, Object>> LMPList = this.LMP.get(refUserEDDId);
//		if (LMPList == null) {
//			LMPList = new ArrayList<HashMap<String, Object>>();
//		}
//
//		LMPList.add(LMPItem);
//		this.LMP.put(refUserEDDId, LMPList);
//		return this;
//	}
//
//	/**
//	 * Get LMP
//	 * 
//	 * @return LMP
//	 **/
//	@JsonProperty("LMP")
//	@ApiModelProperty(value = "")
//	public HashMap<Integer, ArrayList<HashMap<String, Object>>> getLMP() {
//		return LMP;
//	}
//
//	public void setLMP(HashMap<Integer, ArrayList<HashMap<String, Object>>> LMP) {
//		this.LMP = LMP;
//	}

	@Override
	public boolean equals(java.lang.Object o) {
		if (this == o) {
			return true;
		}
		if (o == null || getClass() != o.getClass()) {
			return false;
		}
		EddList eddList = (EddList) o;
		return Objects.equals(this.userId, eddList.userId) && Objects.equals(this.latestEDD, eddList.latestEDD)
				&& Objects.equals(this.eddMapList, eddList.eddMapList); 
	}

	@Override
	public int hashCode() {
		return Objects.hash(userId, eddMapList, latestEDD);
	}

	@Override
	public String toString() {
		StringBuilder sb = new StringBuilder();
		sb.append("class EddList {\n");

		sb.append("    userId: ").append(toIndentedString(userId)).append("\n");
		sb.append("    isEnabled: ").append(toIndentedString(isEnabled)).append("\n");
		sb.append("    eddMapList: ").append(toIndentedString(eddMapList)).append("\n");
//		sb.append("    IVF: ").append(toIndentedString(IVF)).append("\n");
//		sb.append("    LMP: ").append(toIndentedString(LMP)).append("\n");
		sb.append("}");
		return sb.toString();
	}

	/**
	 * Convert the given object to string with each line indented by 4 spaces
	 * (except the first line).
	 */
	private String toIndentedString(java.lang.Object o) {
		if (o == null) {
			return "null";
		}
		return o.toString().replace("\n", "\n    ");
	}
}