# src/gen/java/io/swagger/model/PatientDoctor.java
/*
 * Expected Delivery Date
 * This is an api for getting expected delivery date for a pregnant women.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: aakash.verma@cerner.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package io.swagger.model;

import java.util.Objects;

import com.fasterxml.jackson.annotation.JsonProperty;

import io.swagger.annotations.ApiModelProperty;

/**
 * PatientDoctor
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2019-09-08T11:58:17.132Z")
public class PatientDoctor {

	@JsonProperty("id")
	private Integer id = null;

	@JsonProperty("vistDate")
	private String vistDate = null;

	@JsonProperty("doctor")
	private User doctor = null;

	@JsonProperty("reason")
	private String reason = null;

	public PatientDoctor vistDate(String vistDate) {
		this.vistDate = vistDate;
		return this;
	}

	/**
	 * Get vistDate
	 * 
	 * @return vistDate
	 **/
	@JsonProperty("vistDate")
	@ApiModelProperty(value = "")
	public String getVistDate() {
		return vistDate;
	}

	public void setVistDate(String vistDate) {
		this.vistDate = vistDate;
	}

	public PatientDoctor doctor(User doctor) {
		this.doctor = doctor;
		return this;
	}

	/**
	 * Get doctor
	 * 
	 * @return doctor
	 **/
	@JsonProperty("doctor")
	@ApiModelProperty(value = "")
	public User getDoctor() {
		return doctor;
	}

	public void setDoctor(User doctor) {
		this.doctor = doctor;
	}

	public PatientDoctor reason(String reason) {
		this.reason = reason;
		return this;
	}
	
	public PatientDoctor id(Integer id) {
		this.id = id;
		return this;
	}

	/**
	 * Get reason
	 * 
	 * @return reason
	 **/
	@JsonProperty("reason")
	@ApiModelProperty(value = "")
	public String getReason() {
		return reason;
	}

	public void setReason(String reason) {
		this.reason = reason;
	}
	
	/**
	 * Get id
	 * 
	 * @return id
	 **/
	@JsonProperty("id")
	@ApiModelProperty(value = "")
	public Integer getId() {
		return id;
	}

	public void setId(Integer id) {
		this.id = id;
	}

	@Override
	public boolean equals(java.lang.Object o) {
		if (this == o) {
			return true;
		}
		if (o == null || getClass() != o.getClass()) {
			return false;
		}
		PatientDoctor patientDoctor = (PatientDoctor) o;
		return Objects.equals(this.vistDate, patientDoctor.vistDate)
				&& Objects.equals(this.doctor, patientDoctor.doctor)
				&& Objects.equals(this.reason, patientDoctor.reason)
				&& Objects.equals(this.id, patientDoctor.id);
	}

	@Override
	public int hashCode() {
		return Objects.hash(vistDate, doctor, reason, id);
	}

	@Override
	public String toString() {
		StringBuilder sb = new StringBuilder();
		sb.append("class PatientDoctor {\n");
		sb.append("    id: ").append(toIndentedString(id)).append("\n");
		sb.append("    vistDate: ").append(toIndentedString(vistDate)).append("\n");
		sb.append("    doctor: ").append(toIndentedString(doctor)).append("\n");
		sb.append("    reason: ").append(toIndentedString(reason)).append("\n");
		
		sb.append("}");
		return sb.toString();
	}

	/**
	 * Convert the given object to string with each line indented by 4 spaces
	 * (except the first line).
	 */
	private String toIndentedString(java.lang.Object o) {
		if (o == null) {
			return "null";
		}
		return o.toString().replace("\n", "\n    ");
	}
}